+++ ImplicareCore/src/br/cefetmg/implicare/model/service/VagaManagement.java
@@ -22,7 +22,7 @@ public interface VagaManagement {
     public void insert(Vaga Vaga) throws BusinessException, PersistenceException;
     public boolean update(long CNPJ, int Cod_Cargo, Date Dat_Publicacao,Vaga Vaga) throws BusinessException, PersistenceException;
     public boolean delete(long CNPJ, int Cod_Cargo, Date Dat_Publicacao) throws PersistenceException;
-    public List<Vaga> getVagaCNPJ(int CNPJ) throws PersistenceException;
+    public List<Vaga> getVagaCNPJ(long CNPJ) throws PersistenceException;
     public List<Vaga> getVagaCod_Cargo(List<CargoInteresse> CarInteresse) throws PersistenceException;
     public Vaga getVagaCod(long CNPJ, int Cod_Cargo, Date Dat_Publicacao) throws PersistenceException;
 }
